patient medical management

+++ src/main/java/com/project/Medical/Service/PatientService.java
@@ -4,11 +4,15 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import com.project.Medical.Entity.Address;
 import com.project.Medical.Entity.Patient;
 import com.project.Medical.Exception.PatientNotFoundException;
 import com.project.Medical.Repo.PatientRepo;
+
 
 @Service
 public class PatientService 
@@ -16,9 +20,11 @@ public class PatientService
 	@Autowired
 	private PatientRepo repo;
 	
-	//Save Patients	
+	//Save Patients
+	@Transactional(rollbackFor=Exception.class)
 	public Patient savePatients(Patient patient)
 	{
+		System.out.println("Gender::::"+patient.getGender());
 		return repo.save(patient);
 	}
 
@@ -28,26 +34,29 @@ public class PatientService
 		Patient patient=repo.findById(patientId).orElse(null);
 		
 		if(patient!=null)
-		{
 			return patient;
-		}
 		else
-		{
-			throw new PatientNotFoundException("User not found with id:: "+patientId);
-		}
-		
+			throw new PatientNotFoundException("User not found with id:: "+patientId);		
 	}
 
 	//Delete the Patient
-	public void deletePatient(int patientId) 
+	public void  deletePatient(int patientId) throws PatientNotFoundException 
 	{
-		repo.deleteById(patientId);
+		Patient patient=repo.findById(patientId).orElse(null);
+		if(patient!=null)
+		{
+			repo.deleteById(patientId);
+		}
+		else
+			throw new PatientNotFoundException("User not found with id:: "+patientId);
 	}
 
 	//Update Patient
-	public Patient updatePatient(Patient patient) 
+	public Patient updatePatient(Patient patient)
 	{
-		return repo.save(patient);
+		repo.findById(patient.getPatientId()).orElseThrow();
+		return repo.save(patient); 
+			
 	}
 	
 	//Retrieve all Patients
